+++ lib/ts/recipe/thirdparty/index.ts
@@ -105,6 +105,14 @@ export default class Wrapper {
     static Facebook = thirdPartyProviders.Facebook;
 
     static Apple = thirdPartyProviders.Apple;
+
+    static Discord = thirdPartyProviders.Discord;
+
+    static GoogleWorkspaces = thirdPartyProviders.GoogleWorkspaces;
+
+    static Okta = thirdPartyProviders.Okta;
+
+    static ActiveDirectory = thirdPartyProviders.ActiveDirectory;
 }
 
 export let init = Wrapper.init;
@@ -136,5 +144,13 @@ export let Github = Wrapper.Github;
 export let Facebook = Wrapper.Facebook;
 
 export let Apple = Wrapper.Apple;
+
+export let Discord = Wrapper.Discord;
+
+export let GoogleWorkspaces = Wrapper.GoogleWorkspaces;
+
+export let Okta = Wrapper.Okta;
+
+export let ActiveDirectory = Wrapper.ActiveDirectory;
 
 export type { RecipeInterface, User, APIInterface, APIOptions, TypeProvider };