+++ lib/ts/recipe/thirdparty/providers/index.ts
@@ -2,8 +2,16 @@ import ProviderGoogle from "./google";
 import ProviderFacebook from "./facebook";
 import ProviderGithub from "./github";
 import ProviderApple from "./apple";
+import ProviderDiscord from "./discord";
+import ProviderOkta from "./okta";
+import ProviderGoogleWorkspaces from "./googleWorkspaces";
+import ProviderAD from "./activeDirectory";
 
 export let Google = ProviderGoogle;
 export let Facebook = ProviderFacebook;
 export let Github = ProviderGithub;
 export let Apple = ProviderApple;
+export let Discord = ProviderDiscord;
+export let GoogleWorkspaces = ProviderGoogleWorkspaces;
+export let Okta = ProviderOkta;
+export let ActiveDirectory = ProviderAD;